nosql_db/hello_world/gighunt/pyproject.toml
[tool.poetry]
name = "gighunt"
version = "0.1.0"
description = ""
authors = ["Fayzak <gildem19@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-arango = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
